Move WeakPtrFactory member to the end in ImageFetcher
This forces WeakPtrs to be invalidated before other members' destructors run.
See the bug or the documentation of WeakPtrFactory for more information. I'm
in the process of trying to turn on a clang check for this (see the bug).

BUG=303818
R=sdefresne@chromium.org

Review URL: https://codereview.chromium.org/808113005

Cr-Original-Commit-Position: refs/heads/master@{#311280}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d8c8e56321d4c9135ad7d5015db99fbb0e2ec4c2