Change grit whitelist to a string parameter
This CL changes the grit_whitelist parameter to be a string instead of a
part of grit_defines.
This allow user to override the parameter with another file.

The goal of this is to allow creating bundles with their own whitelists.

At the moment, if you want to create a small executable (i.e. an ios extension)
that only uses 3 strings, you are required to include all the strings
whitelisted by the main chrome bundle.
By allowing to override the whitelist argument (instead of only appending the
argument array, it will allow to specify a specific whitelist for a specific
target.

BUG=456837

Review URL: https://codereview.chromium.org/862043002

Cr-Original-Commit-Position: refs/heads/master@{#315324}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2310fc2b85b31504de87ba22d9d14fc406ed3147