Fix crash occurring during infobar dismissal at BVC deletion.
Calling the InfoBarContainerIOS destructor leads to the removal of infobars via
|InfoBarContainerIOS::PlatformSpecificRemoveInfoBar|, which leads to the BVC
wanting to change the InfoBarContainerIOS's height (in |infoBarContainerStateChanged|)
This leads to a crash because the InfoBarContainerIOS is in the process of
being deleted.

This CL makes the flow abort early.

BUG=448093

Review URL: https://codereview.chromium.org/1038903003

Cr-Original-Commit-Position: refs/heads/master@{#322432}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 695932863ee381ddcb55a61eae9e5f0623df9138