Add MemoryWedge to ios/.
A MemoryWedge is a single chunk of memory allocated. One passes it a size and the wedge is added to the resident memory used by the current app.

Furthermore, it adds an experimental flag for iOS based on a variation param, to drive the size of the wedge that could be inserted for a given run.

BUG=TBD
R=stuartmorgan,asvitkine

Review URL: https://codereview.chromium.org/1107963002

Cr-Original-Commit-Position: refs/heads/master@{#327471}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68537275ee0c46477455e14127fd61c0988e60f9