Adding an IsSynchronized method to BrowsingDataPartition.
This reflects the fact that an ActiveStateManager's active state is out
of sync with its CRWBrowsingDataStore's mode.

WKWebViews can only be created when the BrowsingDataPartition
is fully synchronized.
Added a DCHECK in the web view creation util methods that vend a WKWebView along
with an appropriate comment in the .h

BUG=480654

Review URL: https://codereview.chromium.org/1153143006

Cr-Original-Commit-Position: refs/heads/master@{#332314}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a9523b3e01e810b23be016a4ab1b45c008ddc805