Splitting removeData into 2 protocol methods.
The 2 methods are:
1) removeDataAtStashPath
2) removeDataAtCanonicalPath

The CRWBrowsingDataManager remove method is called on a background thread.
It has no idea if it needs to remove data at the stash path or at the canonical
path. This context needs to be passed in from the CRWBrowsingDataStore.

This information is obtained by looking at the last dispatched stash or restore
operation.
The last dispatched stash or restore operation is annotated with a name so it
can be found at a later time.

BUG=480654

Review URL: https://codereview.chromium.org/1165573003

Cr-Original-Commit-Position: refs/heads/master@{#332338}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b95d6dd1db30050cc942e3b6eae7c513f6ed900e