Fixed unittest by not assigning nil to non-nil properties
The test makes sure that the internals of a WKWebViewConfiguration are not
changed even when the values of the returned object are changed.

This patch changes the internals of the returned WKWebViewConfiguration object
by passing in non-nil values from another WKWebViewConfiguration.

BUG=498841

Review URL: https://codereview.chromium.org/1189803007

Cr-Original-Commit-Position: refs/heads/master@{#334643}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8bdb2e6579a750625406675ca7fe01565a5355e0