Making callbacks be received in the right order
GCD does not guarantee the order in which the completionBlock of
NSOperations are called.
Even though the makeActive/makeInactive NSOperations are serialized
their
completionBlocks are not necessarily serialized.
Added a new NSOperation that calls
|callCompletionHandlerOnMainThread|.
This makes sure that |callCompletionHandlerOnMainThread| is run
before the next NSOperation is run leading to the correct behavior.

Also changed the unittest to check for the fact that the
|completionHandlers| are received in the right order.

BUG=504812

Review URL: https://codereview.chromium.org/1215853002

Cr-Original-Commit-Position: refs/heads/master@{#336604}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd0633ba2d3445a44e324fdaf2448bb8ca2795bc