Check for CGRectZero instead of CGRectNull.
Some definitions:
- null rectangle: CGRectNull = { {inf, inf}, {0, 0} }
- empty rectangle: any rectangle that has a 0 area, ie has a width or a height of 0. That means that CGRectZero = { {0, 0}, {0, 0} }) or CGRectNull are examples of empty rectangles.

Keyboard-change notifications include a begin and an end frame, describing where the keyboard was and where it is going.
When the software keyboard is undocked, some frames can be CGRectZero.

In order to handle this case, the code prior to this CL was checking incorrectly for CGRectNull, when it should have checked CGRectZero. This resulted in the keyboard being incorrectly reported as in hardware mode.

With this CL, the undocking case is properly handled, and the frame checked to be CGRectZero.

BUG=none
R=sdefresne@chromium.org

Review URL: https://codereview.chromium.org/1256553002

Cr-Original-Commit-Position: refs/heads/master@{#340229}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17654d21f686bca112f90aed3d5341b292126dbc