Allowing multiple WeakNSObjectCounters to track the same object
The same key cannot be used to attach the
CRBWeakNSObjectDeallocationObserver from different
WeakNSObjectCounters, since this will
result in the previous associated object getting
deallocated.

This CL uses the address of the shared counter as a key to
attach the CRBWeakNSObjectDeallocationObserver.
The address of the shared counter is unique for every
WeakNSObjectCounter and hence lends itself naturally as
an option to disambiguate WeakNSObjectCounters.

This is a precursor to allow WKWebViews to be tracked
by both the WebViewCounter and
|active_wk_web_view_counter| (inside of web_view_creation_utils).

BUG=480507

Review URL: https://codereview.chromium.org/1259253005

Cr-Original-Commit-Position: refs/heads/master@{#340726}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 84883a4161426d5ea6c6b2a66e23ee2fba85a954