Move ShouldFilterAutofillResult from ChromePasswordManagerClient to PasswordManager
The goal is to put as much of the platform-independent logic to the shared code, and leave the CPMC just a dumb redirector to the embedding system. This is one of many steps.

Also, this CL removes IsPasswordManagementEnabledForCurrentPage() from the client mock in the client unit test. It is dangerous to mix mocking methods with testing the real implementations, and none of the tests actually needs this method mocked.

BUG=474577, 515108

Review URL: https://codereview.chromium.org/1260263002

Cr-Original-Commit-Position: refs/heads/master@{#342332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5410d0bc855495b61ca084a8760d32613f684d01