SdchManager: remove EnableSdchSupport
This change removes support for globally enabling or disabling SDCH. SDCH can
still be enabled or disabled at the URLRequestContext level by creating or
nulling an individual URLRequestContext's SdchManager. The only non-test user of
this method was CrNet, which now only creates an SdchManager if SDCH is to be
enabled.

This change adds a new CrNet integration test to validate that the default
behavior is correct.

BUG=387891

Review URL: https://codereview.chromium.org/1291673003

Cr-Original-Commit-Position: refs/heads/master@{#343889}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 085ac46e3a37b53f963aa836d6277082fb275c01