Seed account info before loading credentials.
This CL ensures that the account info is seeded before reloading
credentials in the iOS token service. This ensures that the GAIA id
and the email are both available when a |OnRefreshTokenAvailable|
notification is sent by the token service.

BUG=521517

Review URL: https://codereview.chromium.org/1296763006

Cr-Original-Commit-Position: refs/heads/master@{#344208}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0a2b1dbc3a89385925b8c9d7472627bb36cf5fe