Remove request group ID from UA with WKWebView
Without the network stack, there is no need to append a request group
ID to the user agent. This removes it in a minimal way; once the
WKWebView-with-network-stack config is removed it'll be easier to
clean up the cruft.

BUG=None

Review URL: https://codereview.chromium.org/1310763002

Cr-Original-Commit-Position: refs/heads/master@{#345078}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b89310841ba07654c97802a5375dd1ae47e3529f