Don't send error messages for touches on iframes
This is a normal case, so there's no reason for it to be treated as an
error. Sending a message causes an expensive JS->ObjC communication (in
fact, quite a few due to the way the touch looks nearby for elements),
which hurts gesture responsiveness.

BUG=532207

Review URL: https://codereview.chromium.org/1344953002

Cr-Original-Commit-Position: refs/heads/master@{#349013}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58d3f29e66dc2168af91953406881949c85b6919