[iOS] Prepare ios/web/ startup for usage by Chrome on iOS.
Add a flag to control the registration of base::AtExitManager as Chrome
on iOS wants to register it earlier to initialise crash reporting support.

Add support for initialisation of base::CommandLine via WebMainParams so
that it can be used for development and tests.

BUG=547725

Review URL: https://codereview.chromium.org/1428603002

Cr-Original-Commit-Position: refs/heads/master@{#356531}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b00d0939aaea3f0048c4f89f746c78067c60cb8