Set Background Color for Web View Content Views
In WKWebView, you can zoom out on web interstitials and cause the
web page underneath to become visible. UIWebView provides a
background underneath the interstitial to prevent this. Since
WKWebView doesn't, this CL sets the background color on the Web
Content View so that a background appears when zooming out on
interstitials on both UI and WKWebView. This CL also moves a call
to self out of CRWWebViewContentView's init method, and disables a
DCHECK that fails as a result.

BUG=545650, 556848

Review URL: https://codereview.chromium.org/1450963002

Cr-Original-Commit-Position: refs/heads/master@{#360177}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 482f3d292623c7709bf27618f3a84a9d822777b1