Check input type in getElementByNameWithParent
The getElementByNameWithParent function is used to find elements for
autofill.
When filling passwords, it is imperative that those are not filled into
plain-text fields. This CL adds a check for the input field type.

The CL is meant to be a quick and easily mergeable fix for a recent
breakage.
Tests and further fixing of the root cause of the breakage will come in a
follow-up.

BUG=569861

Review URL: https://codereview.chromium.org/1524133002

Cr-Original-Commit-Position: refs/heads/master@{#365234}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 968329336ac2207611a7f9fb50883646f7023110