Gating logic only when a view added to the view hierarchy
It does not make sense to change backgroundColor,
addSubviews when the view is being moved
out of the view hierarchy.

Also, |addSubview| is pretty cheap when
the subview is already part of the subviews
so there is no need to gate it specifically.

Review URL: https://codereview.chromium.org/1524343003

Cr-Original-Commit-Position: refs/heads/master@{#365550}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4f0842d9f620ab7a10bbb0366d54dd12adbbf62