Ignore pushState events if there is a pending NavigationItem.
For UIWebView, the click event for an anchor tag will register a load
request.  If a pushState is issued in the onClick event for an anchor
tap, then the pushState will be sent from the previous page, but the
current NavigationItem will correspond with the href of the link.

BUG=567895

Review URL: https://codereview.chromium.org/1538123002

Cr-Original-Commit-Position: refs/heads/master@{#366230}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 028beca87b6de6eda1f0180e724a90c63b3f5f3e