Make the WebController have the size of the window, instead of the size of the screen.
The WebController size was set to the full size of the screen, even when
Chrome was in multitasking mode.
This caused problem when taking snapshots of the tabs: the snapshots
did not match what the tabs really looked like in multitasking.

BUG=572018

Review URL: https://codereview.chromium.org/1541193003

Cr-Original-Commit-Position: refs/heads/master@{#366821}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 890d02d5c5c2b0853b209348d519909da645e411