Fix infinite recursion with CRWWebController's _requireReloadOnDisplay
When _requireReloadOnDisplay is set, the code to re-issue the load can
itself call -view, so it's critical to clear the flag before running
any of the logic, rather than at the end.

BUG=574188
TEST=Go to a page requiring a form submission, then in other tab use
enough memory that the renderer process for the form tab is killed by
the OS; on switching back, the app should not hang and then crash.

Review URL: https://codereview.chromium.org/1562763003

Cr-Original-Commit-Position: refs/heads/master@{#367886}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da39a57167a549a1eeb9b3f79c58a52355f2e440