Add sync settings to IsCreditCardUploadEnabled.
The check for whether Autofill sync is available matches that in autofill::WalletIntegrationAvailable and should have been in the original implementation of this method. Upload should only be available for clients with sync and for users who include Autofill in their chosen sync types.

The second check covers a privacy requirement that we not offer upload to users who have set a passphrase.

BUG=535784

Review URL: https://codereview.chromium.org/1578953005

Cr-Original-Commit-Position: refs/heads/master@{#369763}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 066021ea9e12b9a87e9497a1423a0de52c0a4061