Add FlushStore to CookieStore interface.
This results in consumers no longer having to downcast to
CookieMonster to call the method, and is part of the process
of removing CookieStore::GetCookieMonster entirely.

Also move a lot of the CookieMonster documentation to
the corresponding CookieStore methods, and moved methods
around to have consistent ordering.

BUG=579653
TBR=newt@chromium.org

Review URL: https://codereview.chromium.org/1616443003

Cr-Original-Commit-Position: refs/heads/master@{#370988}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 96f3bab913a1a5a972388023bfc3741839670a74